Fix `vwf create`.
  - Make the vwf root path available to the CLI.
  - Make blank app the default VWF application.
