+++ lib/nodejs/vwfCli.js
@@ -1,4 +1,5 @@
-var fs = require( 'fs-extra' );
+var fs   = require( 'fs-extra' ),
+    path = require( 'path' );
 
 var create = function( applicationPath ) {
 	if ( fs.existsSync ( applicationPath ) ) {
@@ -7,7 +8,7 @@ var create = function( applicationPath ) {
 
 	fs.mkdirpSync( applicationPath );
 
-	fs.copySync( 'support/cli/sample_app/', applicationPath );
+	fs.copySync( path.join( global.vwfRoot, 'support/cli/sample_app/' ), applicationPath );
 
 	return true;
 }
