+++ node-server.js
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
 var server = require( './node_vwf' ),
+    path   = require( 'path' ),
     fs 	   = require( 'fs' ),
     cli    = require( './lib/nodejs/vwfCli.js' );
 
@@ -37,6 +38,24 @@ function printCreateHelp() {
 	console.log("Example:");
 	console.log("  vwf create ~/code/my-new-app");
 }
+
+// Set the VWF directory where VWF files will be served from. Default to
+// "$HOME/.vwf". If not found at $HOME/.vwf, try the current working
+// directory.
+function parseVWFPath () {
+    var home = ( process.env.HOME || process.env.USERPROFILE );
+    var vwfHome = path.join( home, ".vwf" );
+
+    if ( fs.existsSync( path.join( vwfHome, "support/client/lib" ) ) ) {
+        return vwfHome;
+    } else if ( fs.existsSync( path.join( process.cwd(), "support/client/lib" ) ) ) {
+        return process.cwd();
+    } else {
+        consoleError( "Could not find VWF support files." );
+        return false;
+    }
+}
+global.vwfRoot = parseVWFPath();
 
 if ( argv._[0] == 'create' && argv._.length == 1 ) {
 	console.log("'create' requires a PATH to create the new VWF application.");
