+++ node_vwf.js
@@ -83,27 +83,9 @@ function parseApplicationPath () {
     }
 }
 
-// Set the VWF directory where VWF files will be served from. Default to
-// "$HOME/.vwf". If not found at $HOME/.vwf, try the current working
-// directory.
-function parseVWFPath () {
-    var home = ( process.env.HOME || process.env.USERPROFILE );
-    var vwfHome = path.join( home, ".vwf" );
-
-    if ( fs.existsSync( path.join( vwfHome, "support/client/lib" ) ) ) {
-        return vwfHome;
-    } else if ( fs.existsSync( path.join( process.cwd(), "support/client/lib" ) ) ) {
-        return process.cwd();
-    } else {
-        consoleError( "Could not find VWF support files." );
-        return false;
-    }
-}
-
 //Start the VWF server
 function startVWF() {
     global.activeinstances = [];
-    global.vwfRoot = parseVWFPath();
 
     if ( !global.vwfRoot ) {
         // Should not hit this path since the VWF script checks for the existence
