+++ support/cli/sample_app/index.vwf.yaml
@@ -1,76 +1,3 @@
-# Copyright 2012 United States Government, as represented by the Secretary of Defense, Under
-# Secretary of Defense (Personnel & Readiness).
-# 
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-# 
-#   http://www.apache.org/licenses/LICENSE-2.0
-# 
-# Unless required by applicable law or agreed to in writing, software distributed under the License
-# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-# or implied. See the License for the specific language governing permissions and limitations under
-# the License.
-
+# Simplest possible VWF applcation.
+# See http://virtual.wf/web/docs for more information.
 --- 
-extends: http://vwf.example.com/scene.vwf
-properties:
-  navmode: none
-  rotationSpeed: 10
-methods:
-  initializeCamera:
-children:
-  duck:
-    extends: http://vwf.example.com/node3.vwf
-    # # 2: Uncomment to apply the spin-on-click behavior
-    # # ... located in public/types/spin-on-click.vwf.yaml
-    implements:
-    - http://vwf.example.com/spin-on-click.vwf
-    source: duck.dae
-    type: model/vnd.collada+xml
-    properties:
-      rotation: [ 1, 0, 0, 0 ]
-      # spin-on-click configuration parameters
-      spin-rate: 1
-      spin-axis: [ 0, 0, 1 ]
-    # # 1: Uncomment to jump up on each click
-    # scripts:
-    # - |
-    #   this.pointerClick = function() {
-    #     var rot = this.rotation;
-    #     this.rotation = [ rot[0], rot[1] , rot[2], rot[3] + 1 ];
-    #   }
-  sceneCenter:
-    extends: http://vwf.example.com/node3.vwf
-    properties:
-      translation: [ 0, 0, 0 ]
-  dir1:
-    extends: http://vwf.example.com/light.vwf
-    properties:
-      lightType: "point"
-      enable: true
-      distance: 2000
-      intensity: 2
-      color: [ 128, 128, 128 ]
-      translation: [ -400, 400, -900 ]
-  dir2:
-    extends: http://vwf.example.com/light.vwf
-    properties:
-      lightType: "point"
-      enable: true
-      distance: 2000
-      intensity: 2
-      color: [ 128, 128, 128 ]
-      translation: [ 400, 400, 900 ]
-scripts:
-- |
-    this.initialize = function(){
-      this.future( 0 ).initializeCamera();
-    }
-
-    this.initializeCamera = function() {
-      this.camera.transform = [ -1,   0,  0, 0, 
-                                 0,  -1,  0, 0, 
-                                 0,   0,  1, 0, 
-                                16, 600, 90, 1 ];
-      this.camera.far *= 4;
-    }
